+++ src/app/code/community/JetRails/TwoFactor/Model/Adminhtml/Observer.php
@@ -5,7 +5,7 @@
 	 * config.xml file.  This observer fires before any adminhtml action gets executed.  This way if
 	 * a user is not authenticated, we can block access to those actions and redirect to the
 	 * verification page.
-	 * @version         1.0.8
+	 * @version         1.0.9
 	 * @package         JetRails® TwoFactor
 	 * @category        Model
 	 * @author          Rafael Grigorian - JetRails®