+++ src/app/code/community/JetRails/TwoFactor/Model/Adminhtml/Page.php
@@ -3,7 +3,7 @@
 	/**
 	 * Page.php - This class contains the page route constants that are used throughout the module.
 	 * It also contains methods that link the state to the associated route.
-	 * @version         1.0.8
+	 * @version         1.0.9
 	 * @package         JetRails® TwoFactor
 	 * @category        Model
 	 * @author          Rafael Grigorian - JetRails®